removed comment

+++ src/components/auth/index.js
@@ -47,10 +47,10 @@ export default class Component extends React.Component {
             showToast("error", "Username and password is required.", 360000)
             return;
         }
-        // if (this.state.captchaPayload == "" && process.env.NODE_ENV != "development") {
-        //     showToast("error", "Please confirm your not a robot.", 360000)
-        //     return;
-        // }
+        if (this.state.captchaPayload == "" && process.env.NODE_ENV != "development") {
+            showToast("error", "Please confirm your not a robot.", 360000)
+            return;
+        }
         showToast("success", "Logging in, please wait ...", 360000)
         localStorage.setItem('username', Login.username)
         localStorage.setItem('rememberMe', Login.rememberMe)
